feature [javalib]: 0.6.n: javalib Random now inherits from RandomAccess
